// src/types/directory/v3Object.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Directory
 * Directory service
 * OpenAPI spec version: v0.33.5
 */
import type { V3ObjectProperties } from './v3ObjectProperties';

export interface V3Object {
  readonly created_at?: string;
  display_name?: string;
  etag?: string;
  id: string;
  properties?: V3ObjectProperties;
  type: string;
  readonly updated_at?: string;
}